improve reusablity :rocket:

+++ PROD/Speaker.py
@@ -20,7 +20,7 @@ class Speaker:
         tts = gTTS(text=message, lang=self.language, slow=False)
         mp3 = 'TTS_message_pictogram.mp3'
         tts.save(mp3)
-        os.system("omxplayer " + mp3)
+        self.play(mp3)
         logging.info("Play MP3 file with pictogram text")
 
     def celebrate(self, found_pictogram_english_lowercase):
@@ -30,9 +30,18 @@ class Speaker:
         mp3 = "TTS_message_celebrate.mp3"
         tts.save(mp3)
         logging.info("Play MP3 file celebration")
-        os.system("omxplayer " + mp3)
+        self.play(mp3)
         logging.info("Party on!!!")
-        os.system("omxplayer " + self.celebration_sound + " &")
-        time.sleep(20)
-        logging.info("No drinks anymore.. so were gonna leave")
-        os.system("pkill omxplayer")
+        self.play(self.celebration_sound, True, 20, "No drinks anymore.. so were gonna leave")
+
+    def play(self, path_to_file, play_background=False, duration=None, log_msg=None):
+        logging.info("Got file to play: " + path_to_file)
+        logging.info("enjoy..")
+        if play_background:
+            os.system("omxplayer " + path_to_file + " &")
+        else:
+            os.system("omxplayer " + path_to_file)
+        if duration is not None:
+            time.sleep(duration)
+            logging.info(log_msg)
+            os.system("pkill omxplayer")
